fix: 特性探测要用 __cpp_lib_xxx，不能只靠 __has_include
Linux CI 报出的两个真实缺陷：

1. ch05_cpp23 —— clang + libstdc++ 下编译失败：
     error: no template named 'expected' in namespace 'std'
   原因：__has_include(<expected>) 只说明头文件存在，不说明特性可用。
   libstdc++ 的 <expected> 内部还有一层 __cplusplus 版本门，
   门没过时头文件能包含成功但什么都不声明，于是在使用处才报错。
   改成两级判断：__has_include 决定能不能 include，
   __cpp_lib_expected 决定特性是否真的可用。六个特性全部改正，
   并在源码里写清了这两个宏的分工。

2. ch15_stl_source —— GCC 下 std::min 推导失败：
     error: no matching function for call to 'min(long long int&, long int&)'
   chrono 的 count() 在 LP64 Linux 上返回 long，而变量是 long long，
   同宽但不同类型。std::min 是同类型模板，加显式转换。

CI 侧：驱动脚本失败时由 workflow 兜底把输出转成 ::error:: annotation
（不再只依赖脚本自己上报），并给每种多路复用模式分配不同端口，
避免连续 bind 撞上上一轮的 TIME_WAIT。Windows 加一步 python 可用性检查。

+++ src/ch15_stl_source/main.cpp
@@ -1765,7 +1765,10 @@ static void s11_benchmark() {
             sink = static_cast<long long>(fn());
             auto us = std::chrono::duration_cast<std::chrono::microseconds>(
                           std::chrono::steady_clock::now() - t0).count();
-            best = (std::min)(best, us);
+            // 显式转换：chrono 的 count() 在 LP64 Linux 上返回 long，
+            // 而 best 是 long long。两者虽然同宽但是不同类型，
+            // std::min 是同类型模板，推导会失败（GCC 报 no matching function）。
+            best = (std::min)(best, static_cast<long long>(us));
         }
         std::cout << "    " << demo::pad(tag, 40)
                   << std::right << std::setw(8) << best << " us   (校验和 " << sink << ")\n";
